Refactor SDL internal methods to snail
