+++ CMakeLists.txt
@@ -38,8 +38,13 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 
 # Options
-option(WITH_TESTS "Enable building test executable" OFF)
+option(WITH_TESTS "Build test executable" OFF)
 
+if(WITH_TESTS)
+  add_definitions(-DSNAIL_RENDERER_HEADLESS)
+else(WITH_TESTS)
+  add_definitions(-DSNAIL_RENDERER_SDL)
+endif(WITH_TESTS)
 
 # Source files
 set(CORE_SOURCES
@@ -129,11 +134,26 @@ set(CORE_SOURCES
 
 set(CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/cmake")
 
+
 # Executable
-add_library(${PROJECT_NAME} ${CORE_SOURCES})
-add_executable(${PROJECT_NAME}_bin WIN32 MACOSX_BUNDLE main.cpp)
-target_link_libraries(${PROJECT_NAME}_bin ${PROJECT_NAME})
-set_target_properties(${PROJECT_NAME}_bin PROPERTIES OUTPUT_NAME ${PROJECT_NAME})
+if(WITH_TESTS)
+  set (CATCH_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/thirdparty/catch2})
+  add_library(Catch INTERFACE)
+  target_include_directories(Catch INTERFACE ${CATCH_INCLUDE_DIR})
+
+  # Test sources
+  set(TEST_SOURCES
+    tests/tests.cpp
+    tests/serialization.cpp
+    ${CORE_SOURCES}
+    )
+
+  add_executable(${PROJECT_NAME} ${TEST_SOURCES})
+  target_link_libraries(${PROJECT_NAME} Catch)
+else(WITH_TESTS)
+  add_executable(${PROJECT_NAME} WIN32 MACOSX_BUNDLE ${CORE_SOURCES} main.cpp)
+endif(WITH_TESTS)
+
 
 # If a CMake prefix was not provided, assume dependencies are in "./deps"
 if("${CMAKE_PREFIX_PATH}" STREQUAL "")
@@ -273,9 +293,15 @@ endif()
 
 # My libraries
 add_subdirectory(snail)
-target_link_libraries(${PROJECT_NAME} snail)
 
-INSTALL(TARGETS ElonaFoobar_bin
+if(WITH_TESTS)
+  target_link_libraries(${PROJECT_NAME} snail_headless)
+else(WITH_TESTS)
+  target_link_libraries(${PROJECT_NAME} snail)
+endif(WITH_TESTS)
+
+
+INSTALL(TARGETS ElonaFoobar
     BUNDLE DESTINATION . COMPONENT Runtime
     RUNTIME DESTINATION bin COMPONENT Runtime)
 
@@ -297,23 +323,9 @@ INSTALL(CODE "
     fixup_bundle(\"${APPS}\" \"\" \"${DIRS}\")
     " COMPONENT Runtime)
 
-# Testing
 if(WITH_TESTS)
-  set (CATCH_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/thirdparty/catch2})
-  add_library(Catch INTERFACE)
-  target_include_directories(Catch INTERFACE ${CATCH_INCLUDE_DIR})
-
   add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD
                      COMMAND ${CMAKE_COMMAND} -E copy_directory
                      "${CMAKE_CURRENT_SOURCE_DIR}/tests/data"
                      "$<TARGET_FILE_DIR:${PROJECT_NAME}>/tests/data")
-
-  # Test sources
-  set(TEST_SOURCES
-    tests/tests.cpp
-    tests/serialization.cpp
-    )
-  add_executable(test_runner ${TEST_SOURCES})
-  target_link_libraries(test_runner Catch)
-  target_link_libraries(test_runner ${PROJECT_NAME})
-endif()
+endif(WITH_TESTS)
