+++ snail/audio.cpp
@@ -1,9 +1,9 @@
-#include "sdl.hpp"
+#include "audio.hpp"
 
 #if defined(SNAIL_RENDERER_HEADLESS)
-#include "sdl_headless.cpp"
+#include "audio/headless.cpp"
 #elif defined(SNAIL_RENDERER_SDL)
-#include "sdl_impl.cpp"
+#include "audio/sdl.cpp"
 #else
 #error Unsupported renderer
 #endif
