+++ snail/detail/sdl_headless.cpp
@@ -1,0 +1,81 @@
+namespace elona
+{
+namespace snail
+{
+namespace detail
+{
+
+
+
+void enforce_sdl(int result)
+{
+}
+
+
+void enforce_ttf(int result)
+{
+}
+
+
+void enforce_image(int result)
+{
+}
+
+
+
+void enforce_mixer(int result)
+{
+}
+
+
+
+sdl_core::sdl_core()
+{
+}
+
+
+
+sdl_core::~sdl_core()
+{
+}
+
+
+
+sdl_ttf::sdl_ttf()
+{
+}
+
+
+
+sdl_ttf::~sdl_ttf()
+{
+}
+
+
+
+sdl_image::sdl_image()
+{
+}
+
+
+
+sdl_image::~sdl_image()
+{
+}
+
+
+
+sdl_mixer::sdl_mixer()
+{
+}
+
+
+
+sdl_mixer::~sdl_mixer()
+{
+}
+
+
+} // namespace detail
+} // namespace snail
+} // namespace elona
