+++ snail/hsp.cpp
@@ -1,9 +1,9 @@
-#include "sdl.hpp"
+#include "hsp.hpp"
 
 #if defined(SNAIL_RENDERER_HEADLESS)
-#include "sdl_headless.cpp"
+#include "hsp/headless.cpp"
 #elif defined(SNAIL_RENDERER_SDL)
-#include "sdl_impl.cpp"
+#include "hsp/sdl.cpp"
 #else
 #error Unsupported renderer
 #endif
