+++ snail/hsp/headless.cpp
@@ -1,0 +1,109 @@
+namespace elona
+{
+namespace snail
+{
+namespace hsp
+{
+
+int timeGetTime()
+{
+    return 0;
+}
+
+void mes(const std::string&)
+{
+}
+
+void mesbox(std::string&)
+{
+}
+
+void picload(basic_image&, int)
+{
+}
+
+void pos(int, int)
+{
+}
+
+void redraw()
+{
+}
+
+void set_color_mod(int, int, int, int)
+{
+}
+
+void onkey_0()
+{
+}
+
+void await(int)
+{
+}
+
+void boxf(int, int, int, int, const snail::color&)
+{
+}
+
+void boxf(const snail::color&)
+{
+}
+
+void buffer(int, int, int)
+{
+}
+
+void color(int)
+{
+}
+
+void font(int, snail::font_t::style_t, const std::string&)
+{
+}
+
+void gcopy(int, int, int, int, int)
+{
+}
+
+int ginfo(int)
+{
+    return 0;
+}
+
+void gmode(int, int, int, int)
+{
+}
+
+void grotate2(int, int, int, double, int, int)
+{
+}
+
+void grotate(int, int, int, double, int, int)
+{
+}
+
+void gsel(int)
+{
+}
+
+void gzoom(int, int, int, int, int, int, int, bool)
+{
+}
+
+void line(int, int)
+{
+}
+
+void line(int, int, int, int)
+{
+}
+
+void title(const std::string&)
+{
+}
+
+
+} // namespace hsp
+} // namespace snail
+} // namespace elona
